md-cluster: Do not update sb if updatd by other nodes
If an operation such as a device error is recorded in the superblock
by another node, then only the memory structures need to be updated.
So, if this node encounters the device failure at the same (or a little
later) time, the superblock should not be written. Any updates
should be performed in a atomic way at the cluster-node level.

Updates are performed with the tokn lock in hand, so this
guarantees atomicity. If a node reloads the SB, it unsets
MD_CHANGE_DEV and MD_CHANGE_PENDING flags.

Before updating the SB, the node checks
if the update has been performed by another node while we were
waiting for the token lock. If yes, we set the MD_CHANGE_DEVS
and MD_CHANGE_PENDNG bits after we get the token lock (to make sure
we write an update if md_reload_sb() has unset the bits).

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>