md-cluster: Protect communication with a mutex
Communication can happen through multiple threads. It is possible that
one thread steps over another threads sequence.

This also protects from -EBUSY returned by DLM if a node
re-requests for a lock requested earlier.

We use a new flag CLUSTER_MD_COMM_LOCKED which is used to detect
if the mutex is already locked. This is useful for cases where we need to
take and hold the token DLM lock for a while. This bit is set only
under the comm_mutex.

Also, Remove stray/ununsed sb_mutex.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>