md-cluser: make resync_finish only called after pers->sync_request
It is not reasonable that cluster raid to release resync
lock before pers->sync_request is not finished yet.

And the metadata will be changed when node performs resync,
we need to inform other nodes to update metadata, so the
MD_CHANGE_PENDING flag is set before finish resync.

Then metadata_update_finish is move ahead to ensure that
METADATA_UPDATED msg is sent before finish resync, and
metadata_update_start need to be run after "repeat:" label
accordingly.

Signed-off-by: Guoqing Jiang <gqjiang@suse.com>