md-cluster: always setup in-memory bitmap
The in-memory bitmap for raid is allocated on demand,
then for cluster scenario, it is possible that slave
node which received RESYNCING message doesn't have the
in-memory bitmap when master node is perform resyncing,
so we can't make bitmap is match up well among each
nodes.

So for cluster scenario, we need always preserve the
bitmap, and ensure the page will not be freed.

And the next patch is relied on this change since it
keeps sync bitmap among each nodes during resyncing
stage.

Signed-off-by: Guoqing Jiang <gqjiang@suse.com>